Revert "Allow UCI parameters to be double"
This reverts commit 82f7d507eaf83e27a33bf0b433be08d23320b6fe.

Many valid points as following:

1. Almost all eval and search parameters are of type 'int' and whether
   you do the conversion to int inside the tuning algorithm or inside SF,
   seems to not make any difference. (joergoster)

2. Another way to put it is that the objective function we are tuning is
   constant by parts : an input value of 1.3 or 0.6 result in exactly the
   same value of the objective function. Since SPSA converges towards a
   local optimum, our version of SPSA can theorically converge to (almost)
   any value. (Hanamuke)

3. Unless you rewrite the whole evaluation to calculate with doubles instead
   of integers, you are only deceiving yourself with an imagined higher
   precision which simply is not there. The tuner still calculates with the
   correct 'real' numbers, and not with the rounded integer values. (joergoster)

My view is that the cases where you can genuinely tune doubles are extremely rare,
like https://github.com/snicolet/Stockfish/commit/961f8bc91052a159d7be9 and also
in these cases you can use scaling. But in most cases internal parameters are integers
and (main point) you will round to integers anyhow in the final patch to commit.

Secondary and admittely less important points are regardng coding style: UCI protocol
strictly defines parameters as integers and you need a very good reason to not following
this guideline. Moreover to convey to doubles you need a bunch of C-style casts here
and there and still have warnings (as reported by mstembera for MSVC).

But the most sensible point to me is what Joerg very well said: "Unless you rewrite the
whole evaluation to calculate with doubles instead of integers, you are only deceiving
yourself with an imagined higher precision which simply is not there."

No functional change.

+++ src/uci.h
@@ -49,12 +49,12 @@ public:
   Option(OnChange = nullptr);
   Option(bool v, OnChange = nullptr);
   Option(const char* v, OnChange = nullptr);
-  Option(double v, int minv, int maxv, OnChange = nullptr);
+  Option(int v, int minv, int maxv, OnChange = nullptr);
   Option(const char* v, const char* cur, OnChange = nullptr);
 
   Option& operator=(const std::string&);
   void operator<<(const Option&);
-  operator double() const;
+  operator int() const;
   operator std::string() const;
   bool operator==(const char*) const;
